Edited Practice_3

+++ Practice_3/src/Task2.java
@@ -8,7 +8,7 @@ public class Task2 {
     public static ArrayList<String> otherSides(float x){
         ArrayList<String> arr = new ArrayList<>();
         DecimalFormat dec = new DecimalFormat("#.##");
-        arr.add(dec.format((float)(x*Math.sqrt(3))));
+        arr.add(dec.format((x*Math.sqrt(3))));
         arr.add (dec.format(2*x));
 
         return arr;
